backlink_tracker/requirements.txt
Django>=4.0
djangorestframework
celery
redis
selenium
bs4
python-dotenv
